requirements.txt
alembic
asyncpg
email-validator
PyJWT
pydantic
python-dotenv
sanic
SQLAlchemy

